every 1000

+++ Seen_and_Unseen/utilitaires/utils.py
@@ -133,7 +133,7 @@ def MDC(x_hat, x):
     a = tf.math.sqrt(tf.math.reduce_sum(a, axis=2))
     a = tf.math.reduce_sum(a, axis = 1)/n
     tmp = (10/tf.math.log(10))*tf.math.sqrt(2)*STD_DATASET
-    a = tf.math.multiply(a, tmp)
+    a = tf.math.multiply(tf.cast(a, dtype=tf.float64), tf.cast(tmp,dtype=tf.float64))
     return tf.math.reduce_mean(a)
 
 def MDC_1D(x_hat, x):
